# schemas/openapi/control-plane.yaml
openapi: 3.1.0
info:
  title: Automata Control Plane
  version: 0.1.0
  description: |
    HTTP API for the Automata platform.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AutomatonId:
      in: path
      name: automaton_id
      required: true
      schema: { type: string, pattern: '^atm_[a-f0-9]{32}$' }
security:
  - bearerAuth: []
paths:
...
  /v1/automata:
    get:
      summary: List automata
      responses:
        "200": { description: OK }
    post:
      summary: Create an automaton
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAutomatonRequest' }
      responses:
        "201": { description: Created }
  /v1/automata/{automaton_id}:
    parameters: [ { $ref: '#/components/parameters/AutomatonId' } ]
    get:
      summary: Get an automaton
      responses:
        "200": { description: OK }
        "404": { description: Not found }
...
  /v1/treasury/{automaton_id}/fund:
    post:
      summary: Fund an automaton
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/FundRequest' } }
      responses: { "200": { description: OK } }
  /v1/treasury/{automaton_id}/ledger:
    get:
      summary: Read the ledger
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
      responses: { "200": { description: OK } }
...
components:
  schemas:
    CreateAutomatonRequest:
      type: object
      required: [name, genesis_prompt]
      properties:
        name: { type: string }
        genesis_prompt: { type: string }
        parent_id: { type: string, nullable: true }
        initial_balance_micro: { type: integer, minimum: 0 }
        currency: { type: string, default: USDC }
        metadata:
          type: object
          additionalProperties: { type: string }
    FundRequest:
      type: object
      required: [automaton_id, amount_micro]
      properties:
        automaton_id: { type: string }
        amount_micro: { type: integer, minimum: 0 }
        currency: { type: string, default: USDC }
        source: { type: string, default: external }